// client/src/models/album-value/album-value-vote.ts
export interface IAlbumValueVote {
	id: string
	createdAt: Date
	releaseId: string
	rarityGenre: number
	rarityPerformance: number
	formatReleaseScore: number
	integrityGenre: number
	integritySemantic: number
	depthScore: number
	qualityRhymesImages: number
	qualityStructureRhythm: number
	qualityStyleImpl: number
	qualityIndividuality: number
	influenceAuthorPopularity: number
	influenceReleaseAnticip: number
	userId: string
}